+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 ## 0.0.2
 
 * Renamed `nullable` to `read_none` as it was misleading.
+* Can now set defaults as strings.
+* Added `put` method.
+* Renamed `clear` to `reset`.